requirements.txt
# pip config set global.index-url https://pypi.tuna.tsinghua.edu.cn/simple
# pip config set global.trusted-host pypi.tuna.tsinghua.edu.cn
Flask==2.0.1
Werkzeug==2.0.3
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
Flask-Migrate==3.1.0
Flask-JWT-Extended==4.3.1
PyMySQL==1.0.2
requests==2.26.0
python-dotenv==0.19.0
pypinyin>=0.49.0
edge-tts==6.1.9
pytz==2024.1

# Production dependencies
# uWSGI==2.0.19.1  # 仅在 Linux 生产环境中安装

# Testing dependencies
pytest==7.3.1
pytest-flask==1.2.0
pytest-cov==4.1.0

flask
flask-sqlalchemy
flask-migrate
flask-jwt-extended
gunicorn
pymysql
cryptography
redis==4.5.1
